Feature/handle casual disconnect
For outbound, casual-jca is the client with respect to a casual domain that is the server.When the server goes away, for whatever reason, those managed connections that had a connection to that server needs to be destroyed and recreated by the application server.

We did not behave as good citizens in this respect but with this feature we do handle it in a good way.

This resolves Issue #55

Approved-by: Tobias Leo
Approved-by: Chris Kelly