feature/casual-caller-wls-problem (pull request #137)
Noticed on wls only, casual-jca applications JNDI-entries are not found in the tree when casual-caller is deployed ( with correct relative deployment order). We now issue lookups on get if nothing was found during deployment.

Either a users configuration is wrong or we are running on the ever so lovely abondonware appserver wls. Hopefully one day we can drop the support for wls.

Approved-by: Tobias Leo