Feature/casual xa is same rm (#60)
This feature solves issue #59 
From the JTA spec:
"If the target transaction already has another XAResource object participating
in the transaction, the transaction manager invokes the XAResource.isSameRM
method to determine if the specified XAResource represents the same resource
manager instance. This information allows the transaction manager to group the
resource managers that are performing work on behalf of the transaction.
If the XAResource object represents a resource manager instance that has
seen the global transaction before, the transaction manager groups the newly
registered resource together with the previous XAResource object and ensures
that the same resource manager only receives one set of prepare-commit calls for
completing the target global transaction."

We store the domain id of the resource manager we are connected to during the initial handshake.
We match on the domain id, if it is the same domain id then it is the same resource manager.