do not run prepare, commit, rollback on netty threads (#97)
Reverse inbound is in fact a client and as such we do not want prepare/commit/rollback to run on any of nettys networking threads.
This is especially true since in fact an inbound prepare can result in an oubound prepare call via CasualXAResource, that then blocks that thread until an answer is received.

All of the operations are expected to be short running as they are not service calls, service calls stil run on the supplied work manager.