Feature/3.2/bundle gson (#138)
Make sure that gson is not needed in the global module since this polluted all deployed applications with that specific gson implementation.

Two new modules:

casual-json-provider - provides our JsonProvider implementation as well as the dependency on our gson version
casual-api-spi-impl - Module that wraps all our spi modules, currently only casual-json-provider. This is the module that other applications should depend on.